move wais folder, remove unused oauth providers
